Rename setupStorage.ts to loadSnippets.ts
